Completa la autenticacion por rol y permisos
Termina/prueba el registro de usuarios, el login con token, maneja permisos y roles por autorities, agrega validaceiones en los dtos, simplifica repositorios y logica en general, typos

+++ src/main/java/com/ayds/Cloudmerce/repository/PaymentMethodRepository.java
@@ -1,10 +1,11 @@
 package com.ayds.Cloudmerce.repository;
 
-import com.ayds.Cloudmerce.model.entity.PaymentMethodEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import com.ayds.Cloudmerce.model.entity.PaymentMethodEntity;
 
 @Repository
-public interface PaymentMethodRepository extends JpaRepository<PaymentMethodEntity,Integer> {
+public interface PaymentMethodRepository extends JpaRepository<PaymentMethodEntity, Long> {
+
 }
