+++ src/main/java/com/ayds/Cloudmerce/repository/PermissionRepository.java
@@ -1,12 +1,11 @@
 package com.ayds.Cloudmerce.repository;
 
-import com.ayds.Cloudmerce.model.entity.PermissionEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import com.ayds.Cloudmerce.model.entity.PermissionEntity;
+
 @Repository
-public interface PermissionRepository extends JpaRepository<PermissionEntity,Integer> {
-
-
+public interface PermissionRepository extends JpaRepository<PermissionEntity, Long> {
 
 }
